feat: seperated getting bids from status request
